Updated default value in field `date_changed`
Setting default & onupdate to `func.now()` will be get the current time (like calling now() in Postgresql) every time a new record is added or an exiting record is updated, which I guess is the intent. Setting default & onupdate to `dt.datetime.now` will set the default value & onupdate value to the time when the server start and with every new addition or updation of record this preset value would be used.